Replace Supabase ping with keepalive insert
